refactor: reorganise src/Ruvarr into vertical feature slices (#102)
Move files into Downloads/, TvdbSeriesLookup/, TvdbEpisodeLookup/,
ProgramRefreshQueue/, and UnmatchedEpisodes/ feature slices, each with
Commands/, Queries/, Components/, Notifiers/, Extensions/, and Jobs/
sub-folders as applicable. Update all namespaces and using directives.
