+++ src/Ruvarr/Programs/ProgramDetail.razor
@@ -3,8 +3,9 @@
 @implements IAsyncDisposable
 @using Ruvarr.Abstractions
 @using Ruvarr.Contracts
-@using Ruvarr.Downloads
+@using Ruvarr.Downloads.Notifiers
 @using Ruvarr.Downloads.Queries.GetDownloadQueue
+@using Ruvarr.ProgramRefreshQueue.Notifiers
 @using Ruvarr.Programs.Commands.DownloadEpisode
 @using Ruvarr.Programs.Commands.RefreshProgram
 @using Ruvarr.Programs.Queries.GetProgram
