+++ src/Ruvarr/Programs/Programs.razor
@@ -4,6 +4,7 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Ruvarr.Abstractions
 @using Ruvarr.Contracts
+@using Ruvarr.ProgramRefreshQueue.Notifiers
 @using Ruvarr.Programs.Commands.RefreshProgram
 @using Ruvarr.Programs.Queries.GetPrograms
 @inject IServiceScopeFactory ScopeFactory
