+++ src/Ruvarr/UnmatchedEpisodes/Components/UnmatchedEpisodes.razor
@@ -1,4 +1,5 @@
 @page "/unmatched-episodes"
+@namespace Ruvarr.UnmatchedEpisodes.Components
 @inject NavigationManager Navigation
 
 @code {
