+++ tests/Ruvarr.IntegrationTests/Downloads/EpisodeMissingEventHandlerTests.cs
@@ -1,7 +1,7 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 
-using Ruvarr.Downloads;
+using Ruvarr.Downloads.Extensions;
 using Ruvarr.Programs.Domain;
 
 using Shouldly;
