+++ tests/Ruvarr.IntegrationTests/Programs/Episodes/Queries/GetDownloadQueueTests.cs
@@ -5,7 +5,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 
 using Ruvarr.Contracts;
-using Ruvarr.Downloads;
+using Ruvarr.Downloads.Extensions;
 using Ruvarr.Programs.Domain;
 
 using Shouldly;
