+++ tests/Ruvarr.UnitTests/Jobs/RuvProgramRefreshJobTests/SlugRefreshEnqueue.cs
@@ -7,8 +7,9 @@ using NSubstitute;
 using Ruvarr.Infrastructure.Ruv;
 using Ruvarr.Infrastructure.Ruv.Models;
 using Ruvarr.Jobs;
-using Ruvarr.Programs;
+using Ruvarr.ProgramRefreshQueue.Notifiers;
 using Ruvarr.Programs.Domain;
+using Ruvarr.TvdbSeriesLookup.Notifiers;
 using Ruvarr.Testing.Builders;
 
 using Shouldly;
