+++ tests/Ruvarr.UnitTests/Programs/TvdbEpisodeLookupNotifierTests/Enqueue.cs
@@ -1,4 +1,4 @@
-using Ruvarr.Programs;
+using Ruvarr.TvdbEpisodeLookup.Notifiers;
 
 using Shouldly;
