+++ tests/Ruvarr.UnitTests/Programs/TvdbEpisodeLookupNotifierTests/MarkComplete.cs
@@ -1,4 +1,4 @@
-using Ruvarr.Programs;
+using Ruvarr.TvdbEpisodeLookup.Notifiers;
 
 using Shouldly;
